Create Challenge: Which Way is Up?
Use an acceleromoter to determine when your project is facing up with no other motion (e.g. shaking or tilting). When this condition is met, turn on an LED. Otherwise, turn off the LED.

Done in the simulator, so the project is always facing up. Only the the condition of shaking or not shaking the accelerometer can be tested in the simulator.